Refactor: Extract dashboard components into separate files
